Add tasks property to projects
