+++ src/assets/initial-projects.js
@@ -1,0 +1,19 @@
+const INITIAL_PROJECTS = [
+  {
+    name: "Learning React",
+    id: crypto.randomUUID(),
+    date: "Tues Dec 3 2024",
+    description:
+      "Learn React from the ground up. Start with the basics and finish with more advanced knowledge.",
+    tasks: [],
+  },
+  {
+    name: "Mastering React",
+    id: crypto.randomUUID(),
+    date: "Wed Dec 4 2024",
+    description: "These are some really advanced concepts.",
+    tasks: [],
+  },
+];
+
+export default INITIAL_PROJECTS;
